+++ test/honeywell_scanner_test.dart
@@ -1,6 +1,6 @@
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:honeywell_scanner/honeywell_scanner.dart';
+// import 'package:flutter/services.dart';
+// import 'package:flutter_test/flutter_test.dart';
+// import 'package:honeywell_scanner/honeywell_scanner.dart';
 
 void main() {
 //  const MethodChannel channel = MethodChannel('honeywell_scanner');