+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 - **stopScanner** function restores BarcodeScanner to initial properties, this means any custom properties will be replaced with the properties before **startScanner**
 - Honeywell sdk lib AAR updated to v1.00.00.0102
 - Example project updated to show a more detailed how to use.
-- CodeFormatUtils is now an extension of CodeFormat *(breaking change)*
+- CodeFormatUtils is now an extension of CodeFormat **(breaking change)**
+- Changed the way of importing the honeywell.aar library **(breaking change)**
 
 ## 1.0.1+3 (2020-10-10)
 ### Fixed